2018/06/20 DB変更にともなう修正

+++ src/main/java/jp/co/stcinc/api/dto/response/ApplyDetailResponseDto.java
@@ -52,7 +52,7 @@ public class ApplyDetailResponseDto {
 
     // 料金
     @Getter @Setter
-    private Long fare;
+    private Integer fare;
 
     // 備考
     @Getter @Setter
